requirements.txt
numpy
Pillow
gunicorn>=20.0
falcon>=2.0
pytest
